// clients/js-solita-candy-guard/src/generated/types/CandyGuardData.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet';
import { GuardSet, guardSetBeet } from './GuardSet';
import { Group, groupBeet } from './Group';
export type CandyGuardData = {
  default: GuardSet;
  groups: beet.COption<Group[]>;
};

/**
 * @category userTypes
 * @category generated
 */
export const candyGuardDataBeet = new beet.FixableBeetArgsStruct<CandyGuardData>(
  [
    ['default', guardSetBeet],
    ['groups', beet.coption(beet.array(groupBeet))],
  ],
  'CandyGuardData',
);