// clients/js-solita-candy-guard/src/generated/types/Group.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet';
import { GuardSet, guardSetBeet } from './GuardSet';
export type Group = {
  label: string;
  guards: GuardSet;
};

/**
 * @category userTypes
 * @category generated
 */
export const groupBeet = new beet.FixableBeetArgsStruct<Group>(
  [
    ['label', beet.utf8String],
    ['guards', guardSetBeet],
  ],
  'Group',
);